+++ app/Http/Controllers/PinviteController.php
@@ -13,42 +13,53 @@ class PinviteController extends Controller
 
         $pin = new Pinvite();
 
-        $pin->title = $request->get('title');
-        $pin->description = $request->get('description');
-        $pin->thumbnail = $request->get('thumbnail');
-        $pin->latitude = $request->get('latitude');
-        $pin->longitude = $request->get('longitude');
-        $pin->event_time = $request->get('event_time');
+        $pin->title = $request->input('title');
+        $pin->description = $request->input('description');
+        $pin->thumbnail = $request->input('thumbnail');
+        $pin->latitude = $request->input('latitude');
+        $pin->longitude = $request->input('longitude');
+        $pin->event_time = $request->input('event_time');
 
         $pin->entity_id = Entity::create([])->id;
-        $pin->user_id = User::where('api_token', '=', $request->get('api_token'))->first()->id;
+
+        $api_token = $request->header('Authorization');
+        $pin->user_id = User::where('api_token', $api_token)->first()->id;
 
         $pin->save();
 
-        return 1;
+        return $pin;
 
     }
 
-    public function read($entity_id)
+    public function read($pinvite_id)
     {
 
-        $pin = Pinvite::where('entity_id', '=', $entity_id)->first();
+        $pin = Pinvite::find($pinvite_id);
 
         if ($pin == null) {
-            return 0;
+            return response()->json(['error' => 'not found'], 404);
         }
 
         return $pin;
 
     }
 
-    public function update(Request $request, $entity_id)
+    public function update(Request $request, $pinvite_id)
     {
 
-        $pin = Pinvite::where('entity_id', '=', $entity_id)->first();
+        /* Checks if pinvite is there */
+        $pin = Pinvite::find($pinvite_id);
 
         if ($pin == null) {
-            return 0;
+            return response()->json(['error' => 'not found'], 404);
+        }
+
+        /* Checks if pinvite being updated belongs to the user making the
+            request */
+        $api_token = $pin->user->api_token;
+
+        if($api_token != $request->header('Authorization')) {
+            return repsonse()->json(['error' => 'Unauthenticated'], 401);
         }
 
         if ($request->has('title'))
@@ -71,22 +82,45 @@ class PinviteController extends Controller
 
         $pin->update();
 
-        return 1;
+        return $pin;
 
     }
 
-    public function delete($entity_id)
+    public function delete(Request $request, $pinvite_id)
     {
 
-        $pin = Pinvite::where('entity_id', '=', $entity_id)->first();
+        /* Checks if pinvite is there */
+        $pin = Pinpost::find($pinvite_id);
 
         if ($pin == null) {
-            return 0;
+            return response()->json(['error' => 'not found'], 404);
+        }
+
+        /* Checks if pinvite being updated belongs to the user making the
+            request */
+        $api_token = $pin->user->api_token;
+
+        if($api_token != $request->header('Authorization')) {
+            return repsonse()->json(['error' => 'Unauthenticated'], 401);
         }
 
         $pin->delete();
 
-        return 1;
+        return response()->json(['pinvite status' => 'deleted'], 200);
+
+    }
+
+    public function likesCount($pinvite_id)
+    {
+
+        return Pinvite::find($pinvite_id)->entity->likesCount();
+
+    }
+
+    public function commentsCount($pinvite_id)
+    {
+
+        return Pinvite::find($pinvite_id)->entity->commentsCount();
 
     }
